+++ src/components/drawer/index.ts
@@ -1,3 +1,5 @@
 export { default } from './Drawer'
 export type { DrawerProps } from './Drawer'
 export { DrawerSize } from './Drawer'
+export { default as DrawerExampleContent } from './DrawerExampleContent'
+export { default as DrawerFileUpload } from './DrawerFileUpload'