+++ src/components/menu/Menu2.index.ts
@@ -1,0 +1,3 @@
+export { default as Menu2 } from './Menu2'
+export { default as MenuItem2 } from './MenuItem2'
+export type { MenuItem2Variant, MenuItem2Size } from './MenuItem2'