+++ src/components/menu/index.ts
@@ -7,4 +7,10 @@ export { default as MenuSeparator } from './MenuSeparator'
 export { default as MenuShortcut } from './MenuShortcut'
 export { default as MenuDemo } from './MenuDemo'
 
+// Menu 2.0 components
+export { default as Menu2 } from './Menu2'
+export { default as MenuItem2 } from './MenuItem2'
+export { default as Menu2Demo } from './Menu2Demo'
+
 export type { MenuItemState } from './MenuItem'
+export type { MenuItem2Variant, MenuItem2Size } from './MenuItem2'